Random testing (#1)
* Add structure for random testing script

* Add Python script for random testing

Add a python script for performing random search for broken
dependencies. The script takes the amount of runs as a command line
argument and will then proceed to generate and build a random Linux
kernel configuration. It is able to restore existing dependencies from a
file and will add new ones as it finds them.