kasan: add .kunitconfig
Add a .kunitconfig for KASan which is based on the .kunitconfig files
for KCSan and KFence.

However, setting CONFIG_FTRACE=y does not appear to be enough (anymore?)
to satisfy the dependencies for CONFIG_KASAN_TEST, which is why this
.kunitconfig additionally sets CONFIG_STACK_TRACER=y. Alternatively,
CONFIG_BLK_DEV_IO_TRACE=y or any other option that completes the
dependencies of CONFIG_TRACING could be used.

Link: https://lore.kernel.org/all/n37j6cbsogluma25crzruaiq7qcslnjeoroyybsy3vw2cokpcm@mh7r3ocp24cb/T/#u
Signed-off-by: Paul Heidekrüger <paul.heidekrueger@tum.de>